Clear URLs from app.py
