+++ src/murfey/util/models.py
@@ -56,10 +56,6 @@ class ClientInfo(BaseModel):
     id: int
 
 
-class RsyncerSource(BaseModel):
-    source: str
-
-
 class RsyncerInfo(BaseModel):
     source: str
     destination: str
